+++ tiger-proxy/src/main/java/de/gematik/test/tiger/proxy/TigerProxyApplication.java
@@ -4,10 +4,20 @@
 
 package de.gematik.test.tiger.proxy;
 
-import de.gematik.rbellogger.captures.RbelFileReaderCapturer;
+import com.fasterxml.jackson.core.JsonGenerator;
+import com.fasterxml.jackson.core.JsonParser;
+import com.fasterxml.jackson.core.JsonProcessingException;
+import com.fasterxml.jackson.databind.DeserializationContext;
+import com.fasterxml.jackson.databind.JsonDeserializer;
+import com.fasterxml.jackson.databind.JsonSerializer;
+import com.fasterxml.jackson.databind.SerializerProvider;
+import com.fasterxml.jackson.databind.module.SimpleModule;
+import de.gematik.rbellogger.data.RbelElement;
+import de.gematik.rbellogger.data.facet.RbelFacet;
 import de.gematik.test.tiger.common.data.config.tigerProxy.TigerProxyConfiguration;
 import de.gematik.test.tiger.proxy.configuration.ApplicationConfiguration;
 import java.io.File;
+import java.io.IOException;
 import java.util.Base64;
 import java.util.stream.Stream;
 import lombok.Getter;
@@ -32,7 +42,6 @@ public class TigerProxyApplication {
         SpringApplication.run(TigerProxyApplication.class, args);
     }
 
-    @SneakyThrows
     @Bean
     public TigerProxy tigerProxy() {
         if (applicationConfiguration != null) {
@@ -40,5 +49,25 @@ public class TigerProxyApplication {
         } else {
             return new TigerProxy(new TigerProxyConfiguration());
         }
+    }
+
+    @Bean
+    public SimpleModule rbelElementDeserializer() {
+        SimpleModule module = new SimpleModule();
+        module.addSerializer(RbelElement.class, new JsonSerializer<RbelElement>() {
+            @Override
+            public void serialize(RbelElement value, JsonGenerator gen, SerializerProvider serializers)
+                throws IOException {
+                gen.writeStartObject();
+                gen.writeStringField("uuid", value.getUuid());
+                gen.writeArrayFieldStart("facets");
+                for (RbelFacet facet : value.getFacets()) {
+                    gen.writeString(facet.getClass().getSimpleName());
+                }
+                gen.writeEndArray();
+                gen.writeEndObject();
+            }
+        });
+        return module;
     }
 }