+++ tiger-proxy/src/main/resources/rbel.js
@@ -7,6 +7,7 @@
 let lastUuid = "";
 let filterCriterion = "";
 let rootEl;
+let jexlQueryElementUuid = "";
 
 let updateTimeout = 0;
 let updateHandler = null;
@@ -15,6 +16,12 @@ let resetBtn;
 let saveBtn;
 let uploadBtn;
 let quitBtn;
+
+let jexlInspectionResultDiv;
+let jexlInspectionContextDiv;
+let jexlInspectionTreeDiv;
+let jexlInspectionContextParentDiv;
+let jexlInspectionNoContextDiv;
 
 let setFilterCriterionBtn;
 let setFilterCriterionInput;
@@ -51,6 +58,11 @@ document.addEventListener('DOMContentLoaded', function () {
   saveBtn = document.getElementById("saveMsgs");
   uploadBtn = document.getElementById("uploadMsgs");
   quitBtn = document.getElementById("quitProxy");
+  jexlInspectionResultDiv = document.getElementById("jexlResult");
+  jexlInspectionContextDiv = document.getElementById("jexlContext");
+  jexlInspectionTreeDiv = document.getElementById("rbelTree");
+  jexlInspectionContextParentDiv = document.getElementById("contextParent");
+  jexlInspectionNoContextDiv = document.getElementById("jexlNoContext");
 
   setFilterCriterionBtn = document.getElementById("setFilterCriterionBtn");
   setFilterCriterionInput = document.getElementById("setFilterCriterionInput");
@@ -79,6 +91,8 @@ document.addEventListener('DOMContentLoaded', function () {
   setFilterCriterionBtn.addEventListener('click', setFilterCriterion);
   quitBtn.addEventListener('click', quitProxy);
   resetBtn.addEventListener('click', resetMessages);
+  document.getElementById("executeJexlQuery")
+  .addEventListener('click', executeJexlQuery)
   if (tigerProxyUploadUrl === "UNDEFINED") {
     uploadBtn.classList.add("is-hidden");
   } else {
@@ -86,22 +100,22 @@ document.addEventListener('DOMContentLoaded', function () {
   }
 
   document.getElementById("saveHtmlBtn")
-    .addEventListener('click', e => {
-      closeModals();
-      saveHtmlToLocal();
-    });
+  .addEventListener('click', e => {
+    closeModals();
+    saveHtmlToLocal();
+  });
   document.getElementById("saveTrafficBtn")
-    .addEventListener('click', e => {
-      e.preventDefault();
-      closeModals();
-      const a = document.createElement('a');
-      a.style.display = 'none';
-      a.href = "/webui/trafficLog.tgr";
-      a.download = 'trafficLog.tgr';
-      document.body.appendChild(a);
-      a.click();
-      window.URL.revokeObjectURL(url);
-    });
+  .addEventListener('click', e => {
+    e.preventDefault();
+    closeModals();
+    const a = document.createElement('a');
+    a.style.display = 'none';
+    a.href = "/webui/trafficLog.tgr";
+    a.download = 'trafficLog.tgr';
+    document.body.appendChild(a);
+    a.click();
+    window.URL.revokeObjectURL(url);
+  });
   btnOpenRouteModal.addEventListener('click',
       e => {
         btnOpenRouteModal.disabled = true;
@@ -134,7 +148,7 @@ document.addEventListener('DOMContentLoaded', function () {
     })
   });
 
-  document.getElementById("update5").click();
+  document.getElementById("noupdate").click();
 
   btnAddRoute.addEventListener("click", addRoute);
   fieldRouteFrom.addEventListener("keydown", updateAddRouteBtnState);
@@ -165,7 +179,6 @@ function enableModals() {
   // Modals
   let $modalCloses = getAll(
       '.modal-background, .modal-close, .message-header .delete, .modal-card-foot .button');
-
 
   if ($modalCloses.length > 0) {
     $modalCloses.forEach(function ($el) {
@@ -199,6 +212,9 @@ function closeModals() {
   btnOpenRouteModal.disabled = false;
   document.getElementById("routeModalLed").classList.remove("ledactive");
   document.getElementById("routeModalLed").classList.remove("lederror");
+  jexlInspectionResultDiv.classList.add("is-hidden");
+  jexlInspectionContextParentDiv.classList.add("is-hidden");
+  jexlInspectionNoContextDiv.classList.remove("is-hidden");
 }
 
 function enableCardToggles() {
@@ -321,7 +337,7 @@ function resetMessages() {
 function quitProxy() {
   quitBtn.disabled = true;
   const xhttp = new XMLHttpRequest();
-  xhttp.open("GET", "/webui/quit" +  testQuitParam, true);
+  xhttp.open("GET", "/webui/quit" + testQuitParam, true);
   xhttp.onreadystatechange = function () {
     if (this.readyState === 4) {
       if (this.status === 0) {
@@ -373,9 +389,8 @@ function uploadReport() {
       }
     }
   }
-  xhttp.send( encodeURIComponent(document.querySelector("html").innerHTML));
+  xhttp.send(encodeURIComponent(document.querySelector("html").innerHTML));
 }
-
 
 function saveHtmlToLocal() {
   document.querySelector(".navbar").classList.add("is-hidden");
@@ -402,41 +417,123 @@ function saveHtmlToLocal() {
   document.body.removeChild(element);
 }
 
+function addQueryBtn(reqEl) {
+  let titleDiv = getAll(".card-header-title", reqEl)[0].childNodes[0];
+  let titleSpan = getAll("span", titleDiv)[0];
+  let msgUuid = getAll("a", titleDiv)[0].getAttribute("name");
+
+  let queryBtn = document.createElement('a');
+  queryBtn.innerHTML =
+      "<span>Inspect with JEXL</span>";
+  queryBtn.setAttribute("class", "button modal-button is-pulled-right mx-3");
+  queryBtn.setAttribute("data-target", msgUuid);
+  queryBtn.addEventListener("click", function (e) {
+    const $target = document.getElementById("jexlQueryModal");
+    jexlQueryElementUuid = msgUuid;
+    rootEl.classList.add('is-clipped');
+    $target.classList.add('is-active');
+    e.preventDefault();
+    return false;
+  });
+  titleSpan.appendChild(queryBtn);
+}
+
+function executeJexlQuery() {
+  const xhttp = new XMLHttpRequest();
+  let jexlQuery = document.getElementById("jexlQueryInput").value;
+  xhttp.open("GET", "/webui/testJexlQuery"
+      + "?msgUuid=" + jexlQueryElementUuid
+      + "&query=" + encodeURIComponent(jexlQuery),
+      true);
+  xhttp.onreadystatechange = function () {
+    if (this.readyState === 4) {
+      if (this.status === 200) {
+        const response = JSON.parse(this.responseText);
+
+        jexlInspectionTreeDiv.innerHTML =
+            "<h3 class='is-size-4'>Rbel Tree</h3>"
+            + "<pre id='shell'>" + response.rbelTreeHtml + "</pre>";
+        shortenStrings(response);
+        jexlInspectionContextDiv.innerHTML =
+            "<h3 class='is-size-4'>JEXL context</h3>"
+            + "<pre id='json'>"
+            + JSON.stringify(response.messageContext, null, 6)
+            + "</pre>";
+        jexlInspectionContextParentDiv.classList.remove("is-hidden");
+        jexlInspectionNoContextDiv.classList.add("is-hidden");
+        if (response.matchSuccessful) {
+          jexlInspectionResultDiv.innerHTML = "<b>Condition is true:</b>"
+              + "<code class='has-background-dark has-text-danger'>" + jexlQuery+ "</code>";
+          jexlInspectionResultDiv.classList.add("has-background-success");
+          jexlInspectionResultDiv.classList.remove("has-background-primary");
+          jexlInspectionResultDiv.classList.remove("is-hidden");
+        } else {
+          jexlInspectionResultDiv.innerHTML = "<b>Condition is false (or invalid):</b>"
+              + "<code class='has-background-dark has-text-danger'>" + jexlQuery + "</code>";
+          jexlInspectionResultDiv.classList.remove("has-background-success");
+          jexlInspectionResultDiv.classList.add("has-background-primary");
+          jexlInspectionResultDiv.classList.remove("is-hidden");
+        }
+      } else {
+        console.log("ERROR " + this.status + " " + this.responseText);
+      }
+    }
+  }
+  xhttp.send();
+}
+
+function shortenStrings(obj) {
+  for (var property in obj) {
+    if (obj.hasOwnProperty(property)) {
+      if (typeof obj[property] == "object") {
+        shortenStrings(obj[property]);
+      } else {
+        if (typeof obj[property] === 'string' || obj[property]
+            instanceof String) {
+          obj[property] = (obj[property].length > 50) ?
+              obj[property].substr(0, 49) + '...'
+              : obj[property];
+        }
+      }
+    }
+  }
+}
+
+function addSingleMessage(msgMetaData, msgHtmlData) {
+  const listDiv = getAll('.msglist')[0];
+
+  let isRequest = msgMetaData.path;
+  const reqEl = htmlToElement(msgHtmlData);
+  let span = getAll(".msg-sequence", reqEl)[0];
+  span.classList.add("tag", "is-info", "is-light", "mr-3", "is-size-3");
+  span.textContent = msgMetaData.sequenceNumber + 1;
+  addQueryBtn(reqEl);
+  listDiv.appendChild(reqEl);
+
+  if (isRequest) {
+    const menuReq = menuReqHtmlTemplate
+    .replace("${uuid}", msgMetaData.uuid)
+    .replace("${sequence}", msgMetaData.sequenceNumber + 1)
+    .replace("${methodNUrl}", msgMetaData.method + "\n" + msgMetaData.path);
+    document.getElementById("sidebar-menu").appendChild(
+        htmlToElement(menuReq));
+  } else {
+    const menuRes = menuResHtmlTemplate
+    .replace("${uuid}", msgMetaData.uuid)
+    .replace("${sequence}", msgMetaData.sequenceNumber + 1)
+    document.getElementById("sidebar-menu").appendChild(
+        htmlToElement(menuRes));
+  }
+}
+
 function updateMessageList(json) {
   if (json.metaMsgList.length === 0) {
     return;
   }
   let i = 0;
-  const listDiv = getAll('.msglist')[0];
   while (i < json.htmlMsgList.length) {
-    const req = json.metaMsgList[i];
-    if (req.path) {
-      const reqEl = htmlToElement(json.htmlMsgList[i]);
-      let span = getAll(".msg-sequence", reqEl)[0];
-      span.classList.add("tag", "is-info", "is-light", "mr-3", "is-size-3");
-      span.textContent = req.sequenceNumber + 1;
-      listDiv.appendChild(reqEl);
-      const resEl = htmlToElement(json.htmlMsgList[i + 1]);
-      const res = json.metaMsgList[i + 1];
-      span = getAll(".msg-sequence", resEl)[0];
-      span.classList.add("tag", "is-info", "is-light", "mr-3", "is-size-3");
-      span.textContent = res.sequenceNumber + 1;
-      listDiv.appendChild(resEl);
-
-      const menuReq = menuReqHtmlTemplate
-      .replace("${uuid}", req.uuid)
-      .replace("${sequence}", req.sequenceNumber + 1)
-      .replace("${methodNUrl}", req.method + "\n" + req.path);
-      document.getElementById("sidebar-menu").appendChild(
-          htmlToElement(menuReq));
-
-      const menuRes = menuResHtmlTemplate
-      .replace("${uuid}", res.uuid)
-      .replace("${sequence}", res.sequenceNumber + 1)
-      document.getElementById("sidebar-menu").appendChild(
-          htmlToElement(menuRes));
-    }
-    i = i + 2;
+    addSingleMessage(json.metaMsgList[i], json.htmlMsgList[i]);
+    i = i + 1;
   }
   lastUuid = json.metaMsgList[json.metaMsgList.length - 1].uuid;
 
@@ -447,8 +544,10 @@ function updateMessageList(json) {
     const msgListDiv = sidebar.nextElementSibling;
     sidebar.children[sidebar.children.length - 1].scrollIntoView(
         {behaviour: "smooth", block: "end"});
-    msgListDiv.children[msgListDiv.children.length - 1].scrollIntoView(
-        {behaviour: "smooth", block: "end"});
+    if (msgListDiv.children[msgListDiv.children.length - 1] !== undefined) {
+      msgListDiv.children[msgListDiv.children.length - 1].scrollIntoView(
+          {behaviour: "smooth", block: "end"});
+    }
   }
 }
 
@@ -544,5 +643,5 @@ function addRoute() {
 }
 
 function testActivateNoSystemExitOnQuit() {
-  testQuitParam='?noSystemExit=true';
+  testQuitParam = '?noSystemExit=true';
 }